Python/fc3-admin-do-catalogo/requirements.txt
asgiref==3.8.1
Django==4.2.14
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.1
pytest-django==4.8.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
